# api.yaml
openapi: "3.0.3"
info:
  title: Anki Classroom API
  description: |
    Backend API for the Anki Classroom spaced repetition platform.
    All authenticated endpoints require a JWT in the `Authorization: Bearer <token>` header.
    Tokens are obtained via `POST /auth/login` and expire after 24 hours.
    Roles: `admin`, `teacher`, `student`.
  version: "0.1.0"
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudyCard:
      type: object
      properties:
        card_id: { type: integer }
        front: { type: string }
        back: { type: string }
        state: { type: string, enum: [new, learning, review, relearning] }
        due_at: { type: integer, nullable: true }
        stability: { type: number, format: double }
        difficulty: { type: number, format: double }
        reps: { type: integer }
        lapses: { type: integer }
        flag:
          { type: integer, description: "Anki-style card flag 0-7. 0 = no flag." }
        step_index:
          {
            type: integer,
            description: "Current position in the learning/relearning steps list (0-based).",
          }
        deck_title: { type: string }
        predicted_interval:
          type: object
          nullable: true
          description: "Days until next review for each rating (1=Again,2=Hard,3=Good,4=Easy). Only for students."
          properties:
            "1": { type: integer }
            "2": { type: integer }
            "3": { type: integer }
            "4": { type: integer }
    StudySession:
      type: object
      properties:
        deck_id: { type: integer, nullable: true }
        deck_title: { type: string, nullable: true }
        steps:
          type: object
          description: "Deck-level scheduling step intervals (in seconds)."
          properties:
            learning_steps:
              {
                type: array,
                items: { type: integer },
                description: "Learning steps for new cards (seconds), e.g. [60, 600].",
              }
            relearning_steps:
              {
                type: array,
                items: { type: integer },
                description: "Relearning steps for lapsed review cards (seconds), e.g. [600].",
              }
        cards:
          { type: array, items: { $ref: "#/components/schemas/StudyCard" } }
        total_cards: { type: integer }
        reviewed_count: { type: integer }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /decks/{id}/study:
    get:
      tags: [Study]
      summary: Study cards for a deck (must have access via class membership, collaboration, or ownership)
      operationId: deckStudy
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Cards in the deck with scheduling state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Deck is not accessible for study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404": { $ref: "#/components/responses/NotFound" }

  # -----------------------------------------------------------------
  # Note Types
  # -----------------------------------------------------------------
...